fix: 코드 리팩토링 및 로그인/회원가입 페이지 버튼 제거
