+++ src/components/SignUp/Register/StartRegister.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { styled } from "styled-components";
 import { ActiveDot, Dot, SlideCounter } from "../Pagination";
-import { useNavigation } from "../../../hooks/SignIn/Navigation";
+import { useNavigation } from "../../../hooks/SignIn/useNavigation";
 import { StartSubTitle, StartTitle } from "../Title";
 import { motion } from "framer-motion";
